+++ scs-client/C#/SCSSdkClient/SCSSdkTelemetry.cs
@@ -72,7 +72,8 @@ namespace SCSSdkClient {
         public event EventHandler Tollgate;
         public event EventHandler Ferry;
         public event EventHandler Train;
-        public event EventHandler Refuel;
+        public event EventHandler RefuelStart;
+        public event EventHandler RefuelEnd;
 
         public void pause() => _updateTimer.Change(Timeout.Infinite, Timeout.Infinite);
 
@@ -188,9 +189,13 @@ namespace SCSSdkClient {
             if (refuel != scsTelemetry.SpecialEventsValues.Refuel) {
                 refuel = scsTelemetry.SpecialEventsValues.Refuel;
                 if (scsTelemetry.SpecialEventsValues.Refuel) {
-                    Refuel?.Invoke(this, new EventArgs());
+                    RefuelStart?.Invoke(this, new EventArgs());
+                } else {
+                    RefuelEnd?.Invoke(this, new EventArgs());
                 }
             }
+
+   
 
 
             lastTime = time;